requirements.txt
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
